pyproject.toml
[build-system]
build-backend = 'setuptools.build_meta'
requires = [
    "setuptools >= 48",
    "setuptools_scm[toml] >= 4",
    "setuptools_scm_git_archive",
]


[project]

name = "chrysus"
description = "read bank records and pull out some insights"

authors = [
    { name = "Gabe Denton", email = "mason.g.denton@vanderbilt.edu" },
]

requires-python = ">= 3.10, < 3.11"

dynamic = [
    "version", # handled by setuptools_scm, see at the end
]

dependencies = [
    "pydantic >= 2.7.4, < 3",
    "python-dotenv >= 1, < 2",
    "requests > 2.30, < 3",
    "anthropic > 0.45, < 1.0",
    "ratelimit >= 2.2.1, < 3",
    "PyPDF2 >= 3, < 4",
    "pdfplumber > 0.11, < 1",
    "langchain>=0.2.1",
    "langchain-openai>=0.1.7",
    "langchain-anthropic>=0.1.7",
    "langchain-google-genai>=0.1.4",
    "pytesseract",
    "pandas >2 , <3",
    "numpy >2, <3",
    "transformers[torch] >= 4.41, <5",
    "accelerate >= 0.27, <3.0",
    "safetensors >= 0.4, <1",
    "huggingface-hub >= 0.23, <1"


]

[project.optional-dependencies]

dev = [
    "pre-commit",
]

[tool.setuptools.packages.find]
where = ["src", "."]
namespaces = true
include = [
    "chrysus*",
    "tests*",
]

[tool.setuptools.package-data]
"chrysus" = ["version.txt", "**/*.toml"]

[tool.setuptools_scm]
write_to = "src/chrysus/version.txt"
git_describe_command = "git describe --tags --dirty --match 'v*' --abbrev=8"